fix(demo): remarks from review
